Revert "Integration of pre-processing"
