+++ README.md
@@ -31,7 +31,7 @@
 ###### Run:
 
 1. cd into server folder
-2. Run `docker compose --env-file ./app/redis/redis.env  up --build` (this takes a while for the first build). Make sure to use `docker compose` not the (soon to be) deprecated `docker-compose`
+2. Run `docker compose --env-file ./app/redis/redis.env  up --build` (this takes a while for the first build). Make sure to use `docker compose` not the (soon to be) depricated `docker-compose`
 3. Check `localhost:8000/api`in your browser to verify that backend is running
 
 #### Troubleshooting:
