--- scraper/preprocessing/data/Geocatalog_categories.csv
@@ -1,27 +1,0 @@
-category_EN;description_EN;category_DE;description_DE;category_FR;description_FR;category_IT;description_IT
-Society, Culture;
-Health;
-Buildings, Infrastructure;
-Transportation;
-Energy;
-Water and Waste Systems;
-Communication;
-Military, Security;
-Agriculture;
-Economical Activities;
-Geology;
-Soils;
-Natural Hazards;
-Forest, Flora, Fauna;
-Inland Waters;
-Atmosphere, Climatology;
-Environmental Protection;
-Nature Protection;
-Base Maps;
-Land Cover, Land Use;
-Imagery;
-Location;
-Elevation;
-Boundaries;
-Spatial Planning;
-Cadastre, Land Registry;
