--- scraper/preprocessing/requirements.txt
@@ -1,15 +1,0 @@
-bert-extractive-summarizer=0.10.1
-geopandas
-gensim>=4.3.0
-pandas
-PyYAML
-numpy
-nltk
-langdetect
-re2
-rake_nltk>=1.0.6
-spacy>=3.3.1
-spacy-legacy>=3.0.12
-nltk>=3.7
-translators>=5.5.6
-tqdm
