+++ server/app/main.py
@@ -1,4 +1,4 @@
-import os
+
 import json
 import logging
 from typing import Union
@@ -64,12 +64,11 @@ async def startup_event():
     r.ft().config_set("MAXSEARCHRESULTS", "-1" )
 
     global dataframe
-    #WARNING: the path to repo must be adjusted
-    url_github_repo = "https://raw.githubusercontent.com/FHNW-IVGI/Geoharvester/main_preprocessing/scraper/data/"
-    url_geoservices_CH_pkl = os.path.join(url_github_repo, "preprocessed_data.pkl")
-    dataframe = import_pkl_into_dataframe(url_geoservices_CH_pkl)
-    # url_geoservices_CH_csv = "app/tmp/geoservices_CH.csv"
-    # dataframe =  import_csv_into_dataframe(url_geoservices_CH_csv)
+    #url_geoservices_CH_pkl = "app/tmp/rawdata_scraper.pkl" #// We  need to automate this as the csv file is now updated once a week!
+    # dataframe = import_pkl_into_dataframe(url_geoservices_CH_pkl)
+
+    url_geoservices_CH_csv = "app/tmp/geoservices_CH.csv"
+    dataframe =  import_csv_into_dataframe(url_geoservices_CH_csv)
     
     global datajson
     datajson = json.loads(dataframe.to_json(orient='records'))
@@ -140,9 +139,12 @@ async def get_data(query_string: Union[str, None] = None,  service: EnumServiceT
         redis_data = search_redis(redis_query, lang, 0, 40000)
 
         ############################################################################################################################
+        # Testing ranking function from the ranking functions in methods.py
+        # If you want the results from redis you can just set this section as comment
+
         if (query_string != None and len(redis_data.docs) > 0):
-            ranked_results = results_ranking(redis_data.docs, word_list)
-            return paginate(ranked_results)
+            # ranked_results = results_ranking(redis_data.docs, word_list)
+            return paginate(redis_data.docs)
         else:
             pass
         ############################################################################################################################
